requirements.txt
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0
google-generativeai==0.4.1
python-dotenv==1.0.0